Fixed bugs for post feed page
Feed page used to defocus the search bar after a single character had been typed or deleted and swing half the UI off the screen when the search bar was selected. New changes keep original UI in place and shrink the ScrollView appropriately. They also allow for continuous typing in the search bar.