Added LocationService + nearby feeds
Added a class to abstract away all location-related duplicate code. If a screen wants to use location information, it will ask permission if it is not yet granted.

Also updated the main feed and the resources feed to use nearby locations instead of placeholder text boxes. Looks a little empty now tho :(