Massive UI update
Updated UI pages to all fit the same theme and look correct for how far we are in the project. This also means that the dropdown in the menu for making a new post takes its values directly from the appropriate enum and that all pages now have a navbar.